+++ 1_render_clip_comp.py
@@ -23,9 +23,9 @@ client_secret = auth["client_secret"]
 # parameters
 channel = 'sodapoppin'
 max_clips = 30
-date_start = '2023-06-01T00:00:00.00Z'
-date_end = '2023-06-30T00:00:00.00Z'
-min_views_required = 1000
+date_start = '2023-08-01T00:00:00.00Z'
+date_end = '2023-08-31T00:00:00.00Z'
+min_views_required = 500
 get_latest_from_twitch = True
 remove_rendered = True
 clips_to_ignore = [
@@ -37,7 +37,7 @@ clips_to_ignore = [
 # ================================================================
 
 # paths of the cli and data
-path_twitch_cli = path_base + "/thirdparty/Twitch_Downloader_1.53.0/TwitchDownloaderCLI"
+path_twitch_cli = path_base + "/thirdparty/Twitch_Downloader_1.53.2/TwitchDownloaderCLI"
 path_twitch_ffmpeg = path_base + "/thirdparty/ffmpeg-4.3.1-amd64-static/ffmpeg"
 path_twitch_ffprob = path_base + "/thirdparty/ffmpeg-4.3.1-amd64-static/ffprobe"
 path_font = path_base.replace("\\", "/").replace(":", "\\\\:") + "/thirdparty/bebas_neue/BebasNeue-Regular.ttf"
@@ -164,11 +164,16 @@ if get_latest_from_twitch:
             if not utils.terminated_requested and not os.path.exists(file_path):
                 print("\t- download clip: " + str(video['id']))
                 cmd = path_twitch_cli + ' clipdownload' \
-                    + ' --id https://clips.twitch.tv/' + str(video['id']) \
+                    + ' --id ' + str(video['id']) \
+                    + ' --ffmpeg-path "' + path_twitch_ffmpeg + '"' \
                     + ' -o ' + file_path_tmp
-                #subprocess.Popen(cmd, shell=True).wait()
+                # print(cmd)
+                # subprocess.Popen(cmd, shell=True).wait()
                 subprocess.Popen(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL).wait()
-                shutil.move(file_path_tmp, file_path)
+                if os.path.exists(file_path_tmp):
+                    shutil.move(file_path_tmp, file_path)
+                else:
+                    print("\t- VIDEO DOWNLOAD FAILED!!!!")
 
             # CHAT: check if the file exists
             try:
@@ -178,12 +183,16 @@ if get_latest_from_twitch:
                     print("\t- download chat: " + str(video['id']))
                     cmd = path_twitch_cli + ' chatdownload' \
                         + ' --id ' + str(video['id']) \
-                        + ' --embed-images --chat-connections 6' \
+                        + ' --embed-images true --chat-connections 6' \
                         + ' --bttv true --ffz true --stv true' \
                         + ' -o ' + file_path_chat_tmp
+                    # print(cmd)
                     # subprocess.Popen(cmd, shell=True).wait()
                     subprocess.Popen(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL).wait()
-                    shutil.move(file_path_chat_tmp, file_path_chat)
+                    if os.path.exists(file_path_chat_tmp):
+                        shutil.move(file_path_chat_tmp, file_path_chat)
+                    else:
+                        print("\t- not able to download any chat...")
             except Exception as e:
                 print(e)
